Update exchange_calendar_xasx holiday rules
There were some errors in the treatment of Australia Day and ANZAC Day

Prior to 1993 Australia Day was observed on the Monday on or after the 26th of January to provide a long weekend.

This rule was broken in 1993 when the States observed different dates and so the ASX did not close on any of them in that year.

The rule was also broken in the Bi-Centennial year of 1988 when an additional holiday was inserted on Monday 25th 1988 and the Australia Day Holiday was then observed on Tuesday 26th 1988. This produced a long weekend one but breaks the former rule.

In 1994, the States agreed to a common observance for Australia Day according to the present Mondayised rule.

Generally, ANZAC Day is not observed if it falls on a Weekend except it was in 2010. The rule before and after that date was to miss observance of the holiday if it fell on a weekend.

Additional known special closures or holidays since 1-Jan-1970 occurred on:

31-Dec-1984 : ad-hoc since NYE was on a Monday
31-Dec-1990 : ad-hoc since NYE was on a Monday
31-Dec-1999 : ad-hoc closure for Y2K Testing

These dates have been checked against daily time-series back to 1-Jan-1973 and from newspapers where available (e.g. ANZAC Day '93)