Added some missing US President memorial holidays from 1970s and AU/NZ holidays (#101)
* Update README.md

Add project notes on purpose of fork to the README

* Update README.md

* Update README.md

* Update trading_calendar to start from UNIX epoch

Set default start date to 1-Jan-1970 to match MSCI global date range

* Update exchange_calendar_xasx holiday rules

There were some errors in the treatment of Australia Day and ANZAC Day

Prior to 1993 Australia Day was observed on the Monday on or after the 26th of January to provide a long weekend.

This rule was broken in 1993 when the States observed different dates and so the ASX did not close on any of them in that year.

The rule was also broken in the Bi-Centennial year of 1988 when an additional holiday was inserted on Monday 25th 1988 and the Australia Day Holiday was then observed on Tuesday 26th 1988. This produced a long weekend one but breaks the former rule.

In 1994, the States agreed to a common observance for Australia Day according to the present Mondayised rule.

Generally, ANZAC Day is not observed if it falls on a Weekend except it was in 2010. The rule before and after that date was to miss observance of the holiday if it fell on a weekend.

Additional known special closures or holidays since 1-Jan-1970 occurred on:

31-Dec-1984 : ad-hoc since NYE was on a Monday
31-Dec-1990 : ad-hoc since NYE was on a Monday
31-Dec-1999 : ad-hoc closure for Y2K Testing

These dates have been checked against daily time-series back to 1-Jan-1973 and from newspapers where available (e.g. ANZAC Day '93)

* Jevons Global: add some ad-hoc holidays

The NZX has numerous ad-hoc fill-in holidays around X-mas and NYE.

Kiwis like to party :-)

* Jevons Global: extend data back to 1-Jan-1970

There are a number of additional closings to do with Hurricanes, Presidential Elections 1968, 1972 and 1976 and the NY Blackout

HurricaneGloriaClosing
NewYorkBlackout
PresidentialElectionDays

plus additional memorial closings for US Presidents Truman and Johnson.

* Jevons Global: Gloria, NYC Blackout and Elections

To extend data back to 1970 added:
Hurricane Gloria
New York Blackout
Election Days up to 1980
Memorial days for Presidents Truman and Johnson

* Jevons Global: XNYS, XASX, XNZE changes

Completed additions and corrections.

Re-test code and confirm.

* Imported Timestamp to pass tests and added source

# Useful source for the history of New Zealand public holidays:
# Nancy Swarbrick, 'Public holidays', Te Ara - the Encyclopedia of New Zealand
# http://www.TeAra.govt.nz/en/public-holidays/print (accessed 22 December 2019)

* Add additional special holidays to test

Hurricane Gloria, NY Blackout, Presidential Election Days 1972, 1976, 1980 and memorial closings for Presidents Truman and Johnson.

* Lines too long causing CI fail

Test failures on lines to long (over 79 characters)

* Fix evil trailing white space

Who would have thought? Travis should be a lawyer.

* Added import for UTC

remember to import UTC from pytz

* Remove unused import pandas.date_range

Kill pesky white space and unused import

* Fix white space warnings

Blank lines with white space

* White space warning

Repeat offender

* Add blank line before class definition

Never edit in the online tool

* Patch test data for ASX holiday changes

Delete trading day 1990-01-29 Insert trading day 1990-01-26  Reason: Australia Day was then Mondayized to 1990-01-29
Delete trading day 1990-12-31 NYE on Monday
Insert trading day 1993-01-26 In that year no Australia Day closure happened as states observed different days

* Adjust test data for non-trading days added

Following non-trading days were removed from this list:
1990-12-24
1990-12-31
1991-12-27
1992-01-03
1999-12-31

* Revert start date to 1990-01-01

Had changed start date to 1970-01-01 which randomly broke some tests.

It looks like it broke Hong Kong due to lunisolar calendar not extending back that far.

* Remove reference to JG

Just cleaned up comments to remove corporate reference

* Removed ref to JG added source URL

Special NYSE closings document
http://s3.amazonaws.com/armstrongeconomics-wp/2013/07/NYSE-Closings.pdf

* removed comments from README

was messing up completion of my pull request